TrendingSeriesDetailsFragment bug fixed BestsellerFragment bug fix AnimeCatalogFragment bug fix Gutenberg bug fix AuthenticationScreen improvements (small redesign) Small ui improvements Internalization Translation feature added (yandex cloud api)
...

Bug where it wouldnt load trending series details when in production mode

Bug where changing category would add new list content on top of preexisting list

Bug where it would fetch anime api non-stop

Bug where in device it wouldn't load plain-text gutenberg book

Added support for Russian and Portuguese

Translation added for supported languages

+++ movie/screen_movie/trending_movie_details/src/main/kotlin/com/rick/movie/screen_movie/trending_movie_detatils/TrendingMovieDetailsFragment.kt
@@ -212,12 +212,12 @@ class TrendingMovieDetailsFragment : Fragment() {
 
                     imdbChip.text = resources.getString(
                         R.string.movie_screen_movie_trending_movie_details_imdb_rating,
-                        trendingMovie.popularity
+                        trendingMovie.popularity.toInt()
                     )
                     movieDbChip.text =
                         resources.getString(
                             R.string.movie_screen_movie_trending_movie_details_imdb_rating,
-                            trendingMovie.voteAverage
+                            trendingMovie.voteAverage.toInt()
                         )
 
                     similarDetailsAdapter.similarsDiffer
